Proof of concept: parse configs via ObjectMapper
This is by no means a full implementation, and work would be required to
integrate this with our existing system, but it's a very basic sketch of
what the case classes would look like. The advantage is that it would
allow us to delete most of our JSON parsing code, and would make
re-serialization of the config trivial.

Once we settle on a config structure we like, I can work on replacing
our existing parsing with this. The idea is that the various config
classes would be able to generate `Stack.Param` as is currently used to
initialize `Router`s, `Server`s, etc.